Refactoring modules architecture
