+++ seriesbr/bcb/__init__.py
@@ -1,0 +1,3 @@
+from .bcb import get_series, get_metadata, search
+
+__all__ = [get_series, get_metadata, search]
