+++ seriesbr/bcb/json_to_df/search.py
@@ -1,0 +1,10 @@
+import pandas as pd
+
+
+def build_df(json):
+    """Search BCB database through CKAN API."""
+    search_results = json["result"]["results"]
+    assert search_results, "Nothing was found."
+    return pd.DataFrame(search_results).loc[
+        :, ["codigo_sgs", "title", "periodicidade", "unidade_medida"]
+    ]
