+++ seriesbr/ibge/__init__.py
@@ -1,0 +1,3 @@
+from .ibge import get_series, get_metadata, search
+
+__all__ = [get_series, get_metadata, search]
