+++ seriesbr/ibge/json_to_df/__init__.py
@@ -1,0 +1,3 @@
+from . import metadata, search, series
+
+__all__ = [metadata, search, series]
