+++ seriesbr/ibge/url_builders/metadata.py
@@ -1,0 +1,2 @@
+def build_url(table):
+    return f"https://servicodados.ibge.gov.br/api/v3/agregados/{table}/metadados"
