+++ seriesbr/ibge/url_builders/search.py
@@ -1,0 +1,2 @@
+def build_url():
+    return "https://servicodados.ibge.gov.br/api/v3/agregados"
