+++ seriesbr/ipea/__init__.py
@@ -1,0 +1,3 @@
+from .ipea import get_series, get_metadata, search
+
+__all__ = [get_series, get_metadata, search]
