+++ seriesbr/ipea/json_to_df/search.py
@@ -1,0 +1,8 @@
+import pandas as pd
+
+
+def build_df(json):
+    """Search IPEA database."""
+    search_results = json["value"]
+    assert search_results, "Nothing was found."
+    return pd.DataFrame(search_results)
