+++ tests/test_bcb.py
@@ -1,5 +1,3 @@
-import re
-import pytest
 import responses
 import pandas as pd
